Update php-amqplib to renamed project
The videlalvaro/php-amqplib was rename to php-amqplib/php-amqplib
sometime last year ago. The new project offers the same base tags
as the old one, for the most parts it's just a rename.

This needs fixing in syrpc-php since downstream projects can't
currently choose to upgrade to the latest and greatest version of
the lib. It is hoped that this will also fix some of the rabbitmq
issues cropping up when using the lib against old-ish versions of
rabbitmq.